requirements.txt
streamlit==1.48.1
plotly==6.3.0
numpy==1.23.5
scipy==1.10.1
tensorflow-cpu==2.12.0
camb==1.5.2